+++ mail/mail-common/src/mail_scroller/mail_scroller_source/remote_source/remote_conversation_scroller_source.rs
@@ -355,7 +355,7 @@ impl RemoteConversationScrollerSource {
 
         // Save all conversations.
         for conversation in conversations.iter_mut() {
-            conversation.save(&tx).await?
+            conversation.create_or_get_local(&tx).await?
         }
 
         let Some((last, label)) = conversations