+++ core/core-api/src/services/proton/traits.rs
@@ -1,0 +1,12 @@
+pub trait ProtonResponseExt {
+    #[must_use]
+    fn inspect_body(self, f: impl FnOnce(&[u8])) -> Self;
+}
+
+impl ProtonResponseExt for muon::ProtonResponse {
+    fn inspect_body(self, f: impl FnOnce(&[u8])) -> Self {
+        f(self.body());
+
+        self
+    }
+}