+++ mail/mail-uniffi/src/core.rs
@@ -2,6 +2,7 @@ mod crypto;
 pub mod datatypes;
 pub mod human_verification;
 mod keychain;
+mod report_an_issue;
 mod session;
 pub mod validation;