+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-03-xx
 
+### Fixed
+
+ - MailScroller & Prefetcher now stores only non existent data. 
+
 ## [0.59.0] - 2025-03-17
 
 ### Fixed